+++ riana/__main__.py
@@ -2,6 +2,6 @@
 
 """riana.__main__: executed when riana directory is called as script."""
 
-from .main import main
+from riana.cli import main
 
 main()