+++ riana/io/percolator.py
@@ -33,7 +33,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from riana import accmass
+from riana.algorithms import mass_calc as accmass
 from riana.exceptions import DataError
 from riana.records import PSMRecord